last update
i realize that i cant customize the database too much with myth/auth library. So maybe this would be the last time I updated, and change to CodeIgniter 3. Prefer to make all things more customizable, so I can explain to lecturers and friends. Better to make a project that you understand.